Add flag for verbose logging
This change adds a flag verbose_logging
which enables some logfile prints that are disabled by default.
Said prints contribute to most of the size of the logfile,
which can easily become large over time.
By introducing this flag, the size of the logfile in the basic branin
example is reduced from approx. 300K to 16K.