Add C backend for bayesian optimization
Note that this C backend is currently not really adapted for HyperMapper
and is only plugged in through a very crude wrapper.
The C backend is also currently very limited in its functionality.
More functionality can be added later.
Thus, work to be done includes tweaking the backend + wrapper to fit
better into the rest of HyperMapper, and add functionality present in
the other optimizer implementations.