Allowing out-of-search-space samples through cli, resume and default
This change just gives the user more freedom over which data to input. For real world cases this can be useful. In the future more safety features will be added in to make sure that this is what the user wants as well as go allow the same freedom through the non-cli interface.